requirements.txt
flask==3.0.0
pdfplumber==0.10.3
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
nltk==3.8.1
reportlab==4.0.7
language-tool-python==2.8
matplotlib==3.8.2
Pillow==10.1.0